Magnetize
Magnetize is a board game for two players that involves black and white pieces on a rectangular grid; unlike many board games, the board stands up vertically. The objective is to form a square of pieces of your color. Players take turns dropping pieces in chosen columns in the board, where they fall to the lowest possible level. Players can also temporarily turn on a magnet," causing all their pieces to slide as far to the left or right as possible within their rows, and then to cling for some number of turns before falling downwards if they are unsupported.